feat(graph): use mcpnode to solve streamhttpnode
