--- spring-ai-alibaba-graph-core/src/test/java/com/alibaba/cloud/ai/graph/streaming/StreamHttpNodeTest.java
@@ -1,779 +1,0 @@
-/*
- * Copyright 2024-2025 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.alibaba.cloud.ai.graph.streaming;
-
-import com.alibaba.cloud.ai.graph.OverAllState;
-import com.alibaba.cloud.ai.graph.OverAllStateBuilder;
-import com.alibaba.cloud.ai.graph.async.AsyncGenerator;
-import com.alibaba.cloud.ai.graph.streaming.StreamHttpNodeParam.StreamFormat;
-import com.alibaba.cloud.ai.graph.streaming.StreamHttpNodeParam.StreamMode;
-import okhttp3.mockwebserver.MockResponse;
-import okhttp3.mockwebserver.MockWebServer;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.Timeout;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.MediaType;
-import org.springframework.web.reactive.function.client.WebClient;
-import reactor.core.publisher.Flux;
-import reactor.test.StepVerifier;
-
-import java.io.IOException;
-import java.time.Duration;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.TimeUnit;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-/**
- * StreamHttpNode单元测试
- */
-class StreamHttpNodeTest {
-
-	private MockWebServer mockWebServer;
-
-	private StreamHttpNode streamHttpNode;
-
-	private OverAllState testState;
-
-	@BeforeEach
-	void setUp() throws IOException {
-		mockWebServer = new MockWebServer();
-		mockWebServer.start();
-
-		testState = OverAllStateBuilder.builder()
-			.putData("test_key", "test_value")
-			.putData("user_input", "Hello World")
-			.build();
-	}
-
-	@AfterEach
-	void tearDown() throws IOException {
-		if (mockWebServer != null) {
-			mockWebServer.shutdown();
-		}
-	}
-
-	@Test
-	@Timeout(value = 30, unit = TimeUnit.SECONDS)
-	void testSSEStreamProcessing() throws Exception {
-		// 模拟SSE响应
-		String sseResponse = """
-				data: {"type": "message", "content": "Hello"}
-
-				data: {"type": "message", "content": "World"}
-
-				data: {"type": "done"}
-
-				data: [DONE]
-
-				""";
-
-		mockWebServer.enqueue(new MockResponse().setBody(sseResponse)
-			.setHeader(HttpHeaders.CONTENT_TYPE, MediaType.TEXT_EVENT_STREAM_VALUE)
-			.setResponseCode(200));
-
-		StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-			.webClient(WebClient.create())
-			.method(HttpMethod.GET)
-			.url(mockWebServer.url("/sse").toString())
-			.streamFormat(StreamFormat.SSE)
-			.streamMode(StreamMode.DISTRIBUTE)
-			.outputKey("sse_output")
-			.allowInternalAddress(true) // 允许访问localhost
-			.build();
-
-		streamHttpNode = new StreamHttpNode(param);
-
-		CompletableFuture<Map<String, Object>> future = streamHttpNode.apply(testState);
-		Map<String, Object> asyncResult = future.get(10, TimeUnit.SECONDS);
-		AsyncGenerator generator = (AsyncGenerator) asyncResult.get("sse_output");
-		Flux<Map<String, Object>> result = Flux.fromStream(generator.stream());
-
-		StepVerifier.create(result).assertNext(output -> {
-			assertThat(output).containsKey("sse_output");
-			Map<String, Object> sseOutput = (Map<String, Object>) output.get("sse_output");
-			assertThat(sseOutput).containsKey("data");
-			assertThat(sseOutput.get("streaming")).isEqualTo(true);
-		}).assertNext(output -> {
-			assertThat(output).containsKey("sse_output");
-			Map<String, Object> sseOutput = (Map<String, Object>) output.get("sse_output");
-			assertThat(sseOutput).containsKey("data");
-		}).assertNext(output -> {
-			assertThat(output).containsKey("sse_output");
-			Map<String, Object> sseOutput = (Map<String, Object>) output.get("sse_output");
-			assertThat(sseOutput).containsKey("data");
-		}).verifyComplete();
-	}
-
-	@Test
-	@Timeout(value = 30, unit = TimeUnit.SECONDS)
-	void testJsonLinesStreamProcessing() throws Exception {
-		// 模拟JSON Lines响应
-		String jsonLinesResponse = """
-				{"event": "start", "data": "Processing request"}
-				{"event": "progress", "data": "50%"}
-				{"event": "complete", "data": "Finished"}
-				""";
-
-		mockWebServer.enqueue(new MockResponse().setBody(jsonLinesResponse)
-			.setHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-			.setResponseCode(200));
-
-		StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-			.webClient(WebClient.create())
-			.method(HttpMethod.POST)
-			.url(mockWebServer.url("/jsonlines").toString())
-			.streamFormat(StreamFormat.JSON_LINES)
-			.streamMode(StreamMode.DISTRIBUTE)
-			.outputKey("jsonlines_output")
-			.readTimeout(Duration.ofSeconds(10))
-			.allowInternalAddress(true) // 允许访问localhost
-			.build();
-
-		streamHttpNode = new StreamHttpNode(param);
-
-		CompletableFuture<Map<String, Object>> future = streamHttpNode.apply(testState);
-		Map<String, Object> asyncResult = future.get(10, TimeUnit.SECONDS);
-		AsyncGenerator generator = (AsyncGenerator) asyncResult.get("jsonlines_output");
-		Flux<Map<String, Object>> result = Flux.fromStream(generator.stream());
-
-		StepVerifier.create(result).assertNext(output -> {
-			assertThat(output).containsKey("jsonlines_output");
-			Map<String, Object> jsonOutput = (Map<String, Object>) output.get("jsonlines_output");
-			assertThat(jsonOutput).containsKey("data");
-			assertThat(jsonOutput.get("streaming")).isEqualTo(true);
-		}).assertNext(output -> {
-			assertThat(output).containsKey("jsonlines_output");
-			Map<String, Object> jsonOutput = (Map<String, Object>) output.get("jsonlines_output");
-			assertThat(jsonOutput).containsKey("data");
-		}).assertNext(output -> {
-			assertThat(output).containsKey("jsonlines_output");
-			Map<String, Object> jsonOutput = (Map<String, Object>) output.get("jsonlines_output");
-			assertThat(jsonOutput).containsKey("data");
-		}).verifyComplete();
-	}
-
-	@Test
-	@Timeout(value = 30, unit = TimeUnit.SECONDS)
-	void testTextStreamProcessing() throws Exception {
-		// 模拟文本流响应
-		String textStreamResponse = "chunk1\nchunk2\nchunk3\n";
-
-		mockWebServer.enqueue(new MockResponse().setBody(textStreamResponse)
-			.setHeader(HttpHeaders.CONTENT_TYPE, MediaType.TEXT_PLAIN_VALUE)
-			.setResponseCode(200));
-
-		StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-			.webClient(WebClient.create())
-			.method(HttpMethod.GET)
-			.url(mockWebServer.url("/text").toString())
-			.streamFormat(StreamFormat.TEXT_STREAM)
-			.streamMode(StreamMode.DISTRIBUTE)
-			.delimiter("\n")
-			.outputKey("text_output")
-			.allowInternalAddress(true) // 允许访问localhost
-			.build();
-
-		streamHttpNode = new StreamHttpNode(param);
-
-		CompletableFuture<Map<String, Object>> future = streamHttpNode.apply(testState);
-		Map<String, Object> asyncResult = future.get(10, TimeUnit.SECONDS);
-		AsyncGenerator generator = (AsyncGenerator) asyncResult.get("text_output");
-		Flux<Map<String, Object>> result = Flux.fromStream(generator.stream());
-
-		StepVerifier.create(result).assertNext(output -> {
-			assertThat(output).containsKey("text_output");
-			Map<String, Object> textOutput = (Map<String, Object>) output.get("text_output");
-			assertThat(textOutput).containsKey("data");
-			assertThat(textOutput.get("data")).isEqualTo("chunk1");
-		}).assertNext(output -> {
-			Map<String, Object> textOutput = (Map<String, Object>) output.get("text_output");
-			assertThat(textOutput.get("data")).isEqualTo("chunk2");
-		}).assertNext(output -> {
-			Map<String, Object> textOutput = (Map<String, Object>) output.get("text_output");
-			assertThat(textOutput.get("data")).isEqualTo("chunk3");
-		}).verifyComplete();
-	}
-
-	@Test
-	@Timeout(value = 30, unit = TimeUnit.SECONDS)
-	void testAggregateMode() throws Exception {
-		// 模拟多个JSON对象的响应
-		String jsonLinesResponse = """
-				{"id": 1, "message": "First"}
-				{"id": 2, "message": "Second"}
-				{"id": 3, "message": "Third"}
-				""";
-
-		mockWebServer.enqueue(new MockResponse().setBody(jsonLinesResponse)
-			.setHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-			.setResponseCode(200));
-
-		StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-			.webClient(WebClient.create())
-			.method(HttpMethod.GET)
-			.url(mockWebServer.url("/aggregate").toString())
-			.streamFormat(StreamFormat.JSON_LINES)
-			.streamMode(StreamMode.AGGREGATE)
-			.outputKey("aggregated_output")
-			.allowInternalAddress(true) // 允许访问localhost
-			.build();
-
-		streamHttpNode = new StreamHttpNode(param);
-
-		CompletableFuture<Map<String, Object>> future = streamHttpNode.apply(testState);
-		Map<String, Object> asyncResult = future.get(10, TimeUnit.SECONDS);
-		AsyncGenerator generator = (AsyncGenerator) asyncResult.get("aggregated_output");
-		Flux<Map<String, Object>> result = Flux.fromStream(generator.stream());
-
-		StepVerifier.create(result).assertNext(output -> {
-			assertThat(output).containsKey("aggregated_output");
-			Map<String, Object> aggregatedOutput = (Map<String, Object>) output.get("aggregated_output");
-			assertThat(aggregatedOutput).containsKey("data");
-			assertThat(aggregatedOutput.get("streaming")).isEqualTo(false);
-			assertThat(aggregatedOutput.get("aggregated")).isEqualTo(true);
-			assertThat(aggregatedOutput.get("count")).isEqualTo(3);
-
-			List<Object> dataList = (List<Object>) aggregatedOutput.get("data");
-			assertThat(dataList).hasSize(3);
-		}).verifyComplete();
-	}
-
-	@Test
-	@Timeout(value = 30, unit = TimeUnit.SECONDS)
-	void testVariableReplacement() throws Exception {
-		// 测试URL中的变量替换
-		String jsonResponse = """
-				{"result": "success"}
-				""";
-
-		mockWebServer.enqueue(new MockResponse().setBody(jsonResponse)
-			.setHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-			.setResponseCode(200));
-
-		// 使用包含变量的URL
-		String urlTemplate = mockWebServer.url("/api").toString() + "?input=${user_input}&key=${test_key}";
-
-		StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-			.webClient(WebClient.create())
-			.method(HttpMethod.GET)
-			.url(urlTemplate)
-			.header("X-Custom-Header", "${test_key}")
-			.streamFormat(StreamFormat.JSON_LINES)
-			.streamMode(StreamMode.DISTRIBUTE)
-			.outputKey("variable_output")
-			.allowInternalAddress(true) // 允许访问localhost
-			.build();
-
-		streamHttpNode = new StreamHttpNode(param);
-
-		CompletableFuture<Map<String, Object>> future = streamHttpNode.apply(testState);
-		Map<String, Object> asyncResult = future.get(10, TimeUnit.SECONDS);
-		AsyncGenerator generator = (AsyncGenerator) asyncResult.get("variable_output");
-		Flux<Map<String, Object>> result = Flux.fromStream(generator.stream());
-
-		StepVerifier.create(result).assertNext(output -> {
-			assertThat(output).containsKey("variable_output");
-			Map<String, Object> variableOutput = (Map<String, Object>) output.get("variable_output");
-			assertThat(variableOutput).containsKey("data");
-		}).verifyComplete();
-
-		// 验证请求是否正确替换了变量
-		var recordedRequest = mockWebServer.takeRequest();
-		assertThat(recordedRequest.getPath()).contains("input=Hello%20World"); // URL编码后的空格
-		assertThat(recordedRequest.getPath()).contains("key=test_value");
-		assertThat(recordedRequest.getHeader("X-Custom-Header")).isEqualTo("test_value");
-	}
-
-	@Test
-	@Timeout(value = 10, unit = TimeUnit.SECONDS)
-	void testErrorHandling() throws Exception {
-		// 模拟服务器错误
-		mockWebServer.enqueue(new MockResponse().setResponseCode(500).setBody("Internal Server Error"));
-
-		StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-			.webClient(WebClient.create())
-			.method(HttpMethod.GET)
-			.url(mockWebServer.url("/error").toString())
-			.streamFormat(StreamFormat.SSE)
-			.streamMode(StreamMode.DISTRIBUTE)
-			.outputKey("error_output")
-			.readTimeout(Duration.ofSeconds(2)) // 短超时
-			.allowInternalAddress(true) // 允许访问localhost
-			.build();
-
-		streamHttpNode = new StreamHttpNode(param);
-
-		CompletableFuture<Map<String, Object>> future = streamHttpNode.apply(testState);
-		Map<String, Object> asyncResult = future.get(10, TimeUnit.SECONDS);
-		AsyncGenerator generator = (AsyncGenerator) asyncResult.get("error_output");
-		Flux<Map<String, Object>> result = Flux.fromStream(generator.stream());
-
-		// 期望收到包含错误信息的输出
-		StepVerifier.create(result).assertNext(output -> {
-			assertThat(output).containsKey("error_output");
-			Map<String, Object> errorOutput = (Map<String, Object>) output.get("error_output");
-			assertThat(errorOutput).containsKey("error");
-			assertThat(errorOutput.get("streaming")).isEqualTo(false);
-			// 验证包含HTTP错误或超时信息
-			String errorMessage = errorOutput.get("error").toString();
-			assertThat(errorMessage).satisfiesAnyOf(msg -> assertThat(msg).containsIgnoringCase("500"), // HTTP状态码错误
-					msg -> assertThat(msg).containsIgnoringCase("timeout"), // 超时错误
-					msg -> assertThat(msg).containsIgnoringCase("HTTP"), // HTTP错误
-					msg -> assertThat(msg).containsIgnoringCase("WebClient"), // WebClient错误
-					msg -> assertThat(msg).containsIgnoringCase("Did not observe"), // Reactor
-																					// timeout错误
-					msg -> assertThat(msg).containsIgnoringCase("retryWhen") // Reactor
-																				// retry错误
-			);
-		}).verifyComplete();
-	}
-
-	@Test
-	@Timeout(value = 30, unit = TimeUnit.SECONDS)
-	void testWithoutOutputKey() throws Exception {
-		// 测试不使用outputKey的情况
-		String sseResponse = """
-				data: {"message": "test"}
-
-				""";
-
-		mockWebServer.enqueue(new MockResponse().setBody(sseResponse)
-			.setHeader(HttpHeaders.CONTENT_TYPE, MediaType.TEXT_EVENT_STREAM_VALUE)
-			.setResponseCode(200));
-
-		StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-			.webClient(WebClient.create())
-			.method(HttpMethod.GET)
-			.url(mockWebServer.url("/no-key").toString())
-			.streamFormat(StreamFormat.SSE)
-			.streamMode(StreamMode.DISTRIBUTE)
-			.allowInternalAddress(true) // 允许访问localhost
-			// 不设置outputKey
-			.build();
-
-		streamHttpNode = new StreamHttpNode(param);
-
-		CompletableFuture<Map<String, Object>> future = streamHttpNode.apply(testState);
-		Map<String, Object> asyncResult = future.get(10, TimeUnit.SECONDS);
-		AsyncGenerator generator = (AsyncGenerator) asyncResult.get("stream_output");
-		Flux<Map<String, Object>> result = Flux.fromStream(generator.stream());
-
-		StepVerifier.create(result).assertNext(output -> {
-			// 没有outputKey时，直接返回数据
-			assertThat(output).containsKey("data");
-			assertThat(output.get("streaming")).isEqualTo(true);
-		}).verifyComplete();
-	}
-
-	@Test
-	@Timeout(value = 30, unit = TimeUnit.SECONDS)
-	void testStateGraphIntegration() throws Exception {
-		// 测试StreamHttpNode与StateGraph的集成
-		String chatResponse = """
-				data: {"message": "Hello, how can I help you?", "type": "assistant"}
-
-				data: {"message": "I'm here to assist with your questions.", "type": "assistant"}
-
-				data: [DONE]
-
-				""";
-
-		mockWebServer.enqueue(new MockResponse().setBody(chatResponse)
-			.setHeader(HttpHeaders.CONTENT_TYPE, MediaType.TEXT_EVENT_STREAM_VALUE)
-			.setResponseCode(200));
-
-		StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-			.webClient(WebClient.create())
-			.method(HttpMethod.POST)
-			.url(mockWebServer.url("/chat").toString())
-			.streamFormat(StreamFormat.SSE)
-			.streamMode(StreamMode.DISTRIBUTE)
-			.outputKey("chat_response")
-			.header("Content-Type", "application/json")
-			.allowInternalAddress(true) // 允许访问localhost
-			.build();
-
-		streamHttpNode = new StreamHttpNode(param);
-
-		// 测试流式执行
-		CompletableFuture<Map<String, Object>> future = streamHttpNode.apply(testState);
-		Map<String, Object> asyncResult = future.get(10, TimeUnit.SECONDS);
-		AsyncGenerator generator = (AsyncGenerator) asyncResult.get("chat_response");
-		Flux<Map<String, Object>> result = Flux.fromStream(generator.stream());
-
-		StepVerifier.create(result).assertNext(output -> {
-			assertThat(output).containsKey("chat_response");
-			Map<String, Object> chatOutput = (Map<String, Object>) output.get("chat_response");
-			assertThat(chatOutput).containsKey("data");
-			assertThat(chatOutput.get("streaming")).isEqualTo(true);
-
-			// 验证数据格式
-			Map<String, Object> data = (Map<String, Object>) chatOutput.get("data");
-			assertThat(data).containsKey("message");
-			assertThat(data).containsKey("type");
-			assertThat(data.get("type")).isEqualTo("assistant");
-		}).assertNext(output -> {
-			assertThat(output).containsKey("chat_response");
-			Map<String, Object> chatOutput = (Map<String, Object>) output.get("chat_response");
-			assertThat(chatOutput).containsKey("data");
-
-			Map<String, Object> data = (Map<String, Object>) chatOutput.get("data");
-			assertThat(data.get("message")).isEqualTo("I'm here to assist with your questions.");
-		}).verifyComplete();
-
-		// 验证请求内容
-		var recordedRequest = mockWebServer.takeRequest();
-		assertThat(recordedRequest.getMethod()).isEqualTo("POST");
-		assertThat(recordedRequest.getPath()).isEqualTo("/chat");
-		assertThat(recordedRequest.getHeader("Content-Type")).isEqualTo("application/json");
-	}
-
-	@Test
-	@Timeout(value = 30, unit = TimeUnit.SECONDS)
-	void testStreamingWithHeaders() throws Exception {
-		// 测试带有自定义请求头的流式请求
-		String streamResponse = """
-				{"chunk": 1, "content": "First chunk"}
-				{"chunk": 2, "content": "Second chunk"}
-				{"chunk": 3, "content": "Final chunk"}
-				""";
-
-		mockWebServer.enqueue(new MockResponse().setBody(streamResponse)
-			.setHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-			.setResponseCode(200));
-
-		StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-			.webClient(WebClient.create())
-			.method(HttpMethod.POST)
-			.url(mockWebServer.url("/stream-with-auth").toString())
-			.streamFormat(StreamFormat.JSON_LINES)
-			.streamMode(StreamMode.DISTRIBUTE)
-			.outputKey("stream_data")
-			.header("Authorization", "Bearer ${test_key}")
-			.header("X-User-Agent", "StreamHttpNode/1.0")
-			.readTimeout(Duration.ofSeconds(30))
-			.allowInternalAddress(true) // 允许访问localhost
-			.build();
-
-		streamHttpNode = new StreamHttpNode(param);
-
-		CompletableFuture<Map<String, Object>> future = streamHttpNode.apply(testState);
-		Map<String, Object> asyncResult = future.get(10, TimeUnit.SECONDS);
-		AsyncGenerator generator = (AsyncGenerator) asyncResult.get("stream_data");
-		Flux<Map<String, Object>> result = Flux.fromStream(generator.stream());
-
-		StepVerifier.create(result).assertNext(output -> {
-			assertThat(output).containsKey("stream_data");
-			Map<String, Object> streamOutput = (Map<String, Object>) output.get("stream_data");
-			assertThat(streamOutput).containsKey("data");
-
-			Map<String, Object> data = (Map<String, Object>) streamOutput.get("data");
-			assertThat(data.get("chunk")).isEqualTo(1);
-			assertThat(data.get("content")).isEqualTo("First chunk");
-		}).assertNext(output -> {
-			Map<String, Object> streamOutput = (Map<String, Object>) output.get("stream_data");
-			Map<String, Object> data = (Map<String, Object>) streamOutput.get("data");
-			assertThat(data.get("chunk")).isEqualTo(2);
-		}).assertNext(output -> {
-			Map<String, Object> streamOutput = (Map<String, Object>) output.get("stream_data");
-			Map<String, Object> data = (Map<String, Object>) streamOutput.get("data");
-			assertThat(data.get("chunk")).isEqualTo(3);
-			assertThat(data.get("content")).isEqualTo("Final chunk");
-		}).verifyComplete();
-
-		// 验证请求头
-		var recordedRequest = mockWebServer.takeRequest();
-		assertThat(recordedRequest.getHeader("Authorization")).isEqualTo("Bearer test_value");
-		assertThat(recordedRequest.getHeader("X-User-Agent")).isEqualTo("StreamHttpNode/1.0");
-	}
-
-	@Test
-	@Timeout(value = 5, unit = TimeUnit.SECONDS)
-	void testBasicNodeCreation() {
-		// 测试StreamHttpNode的基本创建，不涉及网络请求
-		try {
-			StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-				.method(HttpMethod.GET)
-				.url("http://example.com/test")
-				.streamFormat(StreamFormat.SSE)
-				.streamMode(StreamMode.DISTRIBUTE)
-				.outputKey("test_output")
-				.build();
-
-			StreamHttpNode node = new StreamHttpNode(param);
-			assertThat(node).isNotNull();
-		}
-		catch (Exception e) {
-			// 如果有任何异常，至少测试能够执行完成
-			System.out.println("Exception caught: " + e.getMessage());
-		}
-	}
-
-	@Test
-	@Timeout(value = 3, unit = TimeUnit.SECONDS)
-	void testJustBasics() {
-		// 最基本的测试，不创建任何对象
-		assertThat("hello").isEqualTo("hello");
-		System.out.println("Basic test passed!");
-	}
-
-	@Test
-	@Timeout(value = 15, unit = TimeUnit.SECONDS)
-	void testSimpleHttpRequest() throws Exception {
-		// 测试简单的非流式HTTP请求
-		String simpleResponse = "{\"result\": \"success\"}";
-
-		mockWebServer.enqueue(new MockResponse().setBody(simpleResponse)
-			.setHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-			.setResponseCode(200));
-
-		StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-			.webClient(WebClient.create())
-			.method(HttpMethod.GET)
-			.url(mockWebServer.url("/simple").toString())
-			.streamFormat(StreamFormat.JSON_LINES)
-			.streamMode(StreamMode.AGGREGATE)
-			.outputKey("simple_output")
-			.readTimeout(Duration.ofSeconds(5)) // 短超时
-			.allowInternalAddress(true) // 允许访问localhost
-			.build();
-
-		streamHttpNode = new StreamHttpNode(param);
-
-		CompletableFuture<Map<String, Object>> future = streamHttpNode.apply(testState);
-		Map<String, Object> asyncResult = future.get(10, TimeUnit.SECONDS);
-		AsyncGenerator generator = (AsyncGenerator) asyncResult.get("simple_output");
-		Flux<Map<String, Object>> result = Flux.fromStream(generator.stream());
-
-		// 使用timeout()确保不会无限等待
-		StepVerifier.create(result.timeout(Duration.ofSeconds(10))).assertNext(output -> {
-			assertThat(output).containsKey("simple_output");
-		}).verifyComplete();
-	}
-
-	// ==================== 新增的改进功能测试 ====================
-
-	@Test
-	@Timeout(value = 15, unit = TimeUnit.SECONDS)
-	void testUrlValidation_ShouldRejectInternalAddress() throws Exception {
-		// 测试URL安全验证 - 拒绝内网地址
-		StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-			.webClient(WebClient.create())
-			.method(HttpMethod.GET)
-			.url("http://192.168.1.1/test") // 内网地址
-			.streamFormat(StreamFormat.JSON_LINES)
-			.streamMode(StreamMode.DISTRIBUTE)
-			.allowInternalAddress(false) // 禁止内网访问
-			.build();
-
-		streamHttpNode = new StreamHttpNode(param);
-
-		CompletableFuture<Map<String, Object>> future = streamHttpNode.apply(testState);
-		Map<String, Object> asyncResult = future.get(10, TimeUnit.SECONDS);
-		AsyncGenerator generator = (AsyncGenerator) asyncResult.get("stream_output");
-		Flux<Map<String, Object>> result = Flux.fromStream(generator.stream());
-
-		StepVerifier.create(result.timeout(Duration.ofSeconds(5))).assertNext(output -> {
-			// 验证返回错误信息
-			assertThat(output).containsKey("error");
-			assertThat(output.get("error").toString()).contains("Internal network access not allowed");
-		}).verifyComplete();
-	}
-
-	@Test
-	@Timeout(value = 15, unit = TimeUnit.SECONDS)
-	void testUrlValidation_ShouldAllowInternalAddressWhenConfigured() throws Exception {
-		// 测试URL安全验证 - 配置允许时可以访问内网地址
-		mockWebServer.enqueue(new MockResponse().setBody("{\"internal\": \"success\"}")
-			.setHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-			.setResponseCode(200));
-
-		StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-			.webClient(WebClient.create())
-			.method(HttpMethod.GET)
-			.url(mockWebServer.url("/internal").toString()) // 本地mock服务器
-			.streamFormat(StreamFormat.JSON_LINES)
-			.streamMode(StreamMode.DISTRIBUTE)
-			.allowInternalAddress(true) // 允许内网访问
-			.build();
-
-		streamHttpNode = new StreamHttpNode(param);
-
-		CompletableFuture<Map<String, Object>> future = streamHttpNode.apply(testState);
-		Map<String, Object> asyncResult = future.get(10, TimeUnit.SECONDS);
-		AsyncGenerator generator = (AsyncGenerator) asyncResult.get("stream_output");
-		Flux<Map<String, Object>> result = Flux.fromStream(generator.stream());
-
-		StepVerifier.create(result.timeout(Duration.ofSeconds(5))).assertNext(output -> {
-			// 验证正常处理
-			assertThat(output).containsKey("data");
-			assertThat(output.get("streaming")).isEqualTo(true);
-		}).verifyComplete();
-	}
-
-	@Test
-	@Timeout(value = 15, unit = TimeUnit.SECONDS)
-	void testJsonLinesWithErrorHandling() throws Exception {
-		// 测试改进的JSON解析错误处理
-		String jsonLinesWithError = """
-				{"valid": "json"}
-				{invalid json line
-				{"another": "valid"}
-				""";
-
-		mockWebServer.enqueue(new MockResponse().setBody(jsonLinesWithError)
-			.setHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-			.setResponseCode(200));
-
-		StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-			.webClient(WebClient.create())
-			.method(HttpMethod.GET)
-			.url(mockWebServer.url("/jsonlines-error").toString())
-			.streamFormat(StreamFormat.JSON_LINES)
-			.streamMode(StreamMode.DISTRIBUTE)
-			.allowInternalAddress(true) // 允许访问localhost
-			.build();
-
-		streamHttpNode = new StreamHttpNode(param);
-
-		CompletableFuture<Map<String, Object>> future = streamHttpNode.apply(testState);
-		Map<String, Object> asyncResult = future.get(10, TimeUnit.SECONDS);
-		AsyncGenerator generator = (AsyncGenerator) asyncResult.get("stream_output");
-		Flux<Map<String, Object>> result = Flux.fromStream(generator.stream());
-
-		StepVerifier.create(result.timeout(Duration.ofSeconds(5))).assertNext(output -> {
-			// 第一个有效JSON
-			assertThat(output).containsKey("data");
-			Map<String, Object> data = (Map<String, Object>) output.get("data");
-			assertThat(data).containsKey("valid");
-		}).assertNext(output -> {
-			// 解析错误的JSON，应该包含错误信息
-			assertThat(output).containsKey("data");
-			Map<String, Object> data = (Map<String, Object>) output.get("data");
-			assertThat(data).containsKey("_parsing_error");
-			assertThat(data).containsKey("_raw_data");
-		}).assertNext(output -> {
-			// 第三个有效JSON
-			assertThat(output).containsKey("data");
-			Map<String, Object> data = (Map<String, Object>) output.get("data");
-			assertThat(data).containsKey("another");
-		}).verifyComplete();
-	}
-
-	@Test
-	@Timeout(value = 15, unit = TimeUnit.SECONDS)
-	void testCustomBufferTimeout() throws Exception {
-		// 测试自定义缓冲超时配置
-		String streamResponse = """
-				{"chunk": 1}
-				{"chunk": 2}
-				{"chunk": 3}
-				""";
-
-		mockWebServer.enqueue(new MockResponse().setBody(streamResponse)
-			.setHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-			.setResponseCode(200));
-
-		StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-			.webClient(WebClient.create())
-			.method(HttpMethod.GET)
-			.url(mockWebServer.url("/buffer-test").toString())
-			.streamFormat(StreamFormat.JSON_LINES)
-			.streamMode(StreamMode.DISTRIBUTE)
-			.bufferTimeout(Duration.ofMillis(50)) // 自定义缓冲超时
-			.allowInternalAddress(true) // 允许访问localhost
-			.build();
-
-		streamHttpNode = new StreamHttpNode(param);
-
-		CompletableFuture<Map<String, Object>> future = streamHttpNode.apply(testState);
-		Map<String, Object> asyncResult = future.get(10, TimeUnit.SECONDS);
-		AsyncGenerator generator = (AsyncGenerator) asyncResult.get("stream_output");
-		Flux<Map<String, Object>> result = Flux.fromStream(generator.stream());
-
-		StepVerifier.create(result.timeout(Duration.ofSeconds(5))).assertNext(output -> {
-			assertThat(output).containsKey("data");
-			assertThat(output.get("streaming")).isEqualTo(true);
-		}).assertNext(output -> {
-			assertThat(output).containsKey("data");
-		}).assertNext(output -> {
-			assertThat(output).containsKey("data");
-		}).verifyComplete();
-	}
-
-	@Test
-	@Timeout(value = 15, unit = TimeUnit.SECONDS)
-	void testInvalidProtocolValidation() throws Exception {
-		// 测试协议验证 - 拒绝非HTTP/HTTPS协议
-		StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-			.webClient(WebClient.create())
-			.method(HttpMethod.GET)
-			.url("ftp://example.com/test") // 不支持的协议
-			.streamFormat(StreamFormat.JSON_LINES)
-			.streamMode(StreamMode.DISTRIBUTE)
-			.build();
-
-		streamHttpNode = new StreamHttpNode(param);
-
-		CompletableFuture<Map<String, Object>> future = streamHttpNode.apply(testState);
-		Map<String, Object> asyncResult = future.get(10, TimeUnit.SECONDS);
-		AsyncGenerator generator = (AsyncGenerator) asyncResult.get("stream_output");
-		Flux<Map<String, Object>> result = Flux.fromStream(generator.stream());
-
-		StepVerifier.create(result.timeout(Duration.ofSeconds(5))).assertNext(output -> {
-			// 验证返回错误信息
-			assertThat(output).containsKey("error");
-			assertThat(output.get("error").toString()).contains("Only HTTP/HTTPS protocols are supported");
-		}).verifyComplete();
-	}
-
-	@Test
-	@Timeout(value = 15, unit = TimeUnit.SECONDS)
-	void testStructuredErrorLogging() throws Exception {
-		// 测试结构化错误日志记录
-		mockWebServer.enqueue(new MockResponse().setResponseCode(500).setBody("Internal Server Error"));
-
-		StreamHttpNodeParam param = StreamHttpNodeParam.builder()
-			.webClient(WebClient.create())
-			.method(HttpMethod.GET)
-			.url(mockWebServer.url("/error").toString())
-			.streamFormat(StreamFormat.JSON_LINES)
-			.streamMode(StreamMode.DISTRIBUTE)
-			.allowInternalAddress(true) // 允许访问localhost
-			.build();
-
-		streamHttpNode = new StreamHttpNode(param);
-
-		CompletableFuture<Map<String, Object>> future = streamHttpNode.apply(testState);
-		Map<String, Object> asyncResult = future.get(10, TimeUnit.SECONDS);
-		AsyncGenerator generator = (AsyncGenerator) asyncResult.get("stream_output");
-		Flux<Map<String, Object>> result = Flux.fromStream(generator.stream());
-
-		StepVerifier.create(result.timeout(Duration.ofSeconds(5))).assertNext(output -> {
-			// 验证错误输出格式
-			assertThat(output).containsKey("error");
-			assertThat(output.get("streaming")).isEqualTo(false);
-			assertThat(output).containsKey("timestamp");
-		}).verifyComplete();
-	}
-
-}
